requirements.txt
requests
openai
markdownify
PyMuPDF
camelot-py
